requirements.txt
pywebview
qtpy
flask
